change endpoint name, fix label  attribute, fix styling
Commit changes endpoint name to /comments to follow resource naming best practices.
The  attribute in the labels are corrected to point to the input uid=1000(mafeponce) gid=1000(mafeponce) groups=1000(mafeponce),4(adm),27(sudo),999(docker).
Uses  to generate a line break instead of  for clarity.
Makes use of the  type input in the comment form.