build comment boxes and style them
Commit makes use of a new Comment java class to store comment's information separately instead of using a string. A helper function was added to the getComments() function to build comment boxes inside the comment history that show each comment's submitter name and the comment's timestamp. If an email was provided, the name in the comment links to the email.