add basic authentication implementation
Comment form doesn't load unless user has logged in. If user isn't logged in, they are provided a link to log in and then redirected to the page where they'll now be able to see the comment box to submit. The email field was removed from the comment form since now this can be retrieved directly from the login credentials. A logout link is provided as well.