change doDelete implementation, move data-user-id, fix other issues
The doDelete servlet method implementation was changed to make it more efficient by filtering the queries before preparing the query results, and the redirect to home was removed to fix the Error 405. The data- attribute that stores the user's id was moved to the body element. The email link previously included in user's past comments has been removed, since there is no longer an option to opt out of including you email. The login and logout wording above the comment history has been changed so that the text inside the hyperlink is exactly the function of the link.