add details to README.md, change deleteMarker implementation, add java helper delete methods
Details regarding what the config.js file should include were added to the README.md file, so others know to include their own Google Maps API key in there. The deleteMarker() function now handles both the front-end removal of markers and the back-end. The back-end removal of markers is now implemented using the marker's id instead of its lat and lng values, to avoid situations where querying the marker to delete results in multiple markers when there exists multiple markers with the same coordinates. A helper class and helper methods were added in java to assist in the deletion of comments and markers, avoiding some code repetition and clutter.